Add start of webauthn library.
Okay, so there's two reasons I went with go. One is because the rest of
the software I'm writing lately is go because it deals with containers.
The other reason is because of this library which appears to have pretty
good logic and structure around the server-side of the webauthn flow.

At this point I've only added the basics so that the first instructions
of the Readme compile, nothing works yet, but that's a good enough
checkpoint to be useful.