Save sessions, send registration completion data.
There are quite a few things here that need to work together.

I'm pretty confident the example psuedocode was missing a call to save
the session in beginRegistration. I've added that. We also need a way to
keep track of sessions. The data itself does not appear to have a
reasonable key other than the user ID which is stored as an array of
bytes. Instead of dealing with that directly I'm augmenting the
psuedocode so that it takes the user ID directly.

This means I need a mapping of user ID to session, and that means I need
to initialize my datastore differently to avoid a nil map.

I added a URL-encoded-base64 creation function to JavaScript land for
encoding the arrays of bytes I've got on the client. I also switched
registration completion to POST since it includes a bunch of data.