Working demo.
Okay, this commit is a HUGE mess. It took a really long time in many
broken sessions to get it. I relied heavily on
https://github.com/NHAS/webauthn-example.git
and you can tell because I copied a number of function over wholesale
even though they are redundant.

Does this make me look like a Go/Gin/WebAuthN newb? Yes. It should. I
am.

I'm hoping to get it cleaned up later, but for now the fact that it
works is more valuable than the shame in its ugliness.