Working gob-based serialization of user database
This took a really long time to figure out. Turns out that Go doesn't
complain if you attempt to serialize a pointer. It doesn't fail
compilation and it doesn't emit an error, it just doesn't serialize it
either. There's probably esoteric reasons for this, but it was annoying
as a new Go programmer.

I had to make the fields on User public (first letter capitalization) in
order for gob to be able to serialize it as well.

Still, at this point, we've got the ability to persist the user database
across process restarts.